活动数据回显增加分裂优惠卷信息

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/market/coopen/CoopenListVo.java
@@ -27,7 +27,7 @@ public class CoopenListVo {
     /**
      * 是否永久有效
      */
-    private Integer    isForever;
+    private Byte    isForever;
     private Timestamp startDate;
     private Timestamp endDate;
     private Byte status;
